refactor(project): move ci safety helpers to shared packages
