+++ cmd/project/ci.go
@@ -3,7 +3,6 @@ package project
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"os"
 	"os/exec"
@@ -15,10 +14,12 @@ import (
 
 	"github.com/shopware/shopware-cli/internal/ci"
 	"github.com/shopware/shopware-cli/internal/extension"
+	internalgit "github.com/shopware/shopware-cli/internal/git"
 	"github.com/shopware/shopware-cli/internal/mjml"
 	"github.com/shopware/shopware-cli/internal/packagist"
 	"github.com/shopware/shopware-cli/internal/phpexec"
 	"github.com/shopware/shopware-cli/internal/shop"
+	"github.com/shopware/shopware-cli/internal/system"
 	"github.com/shopware/shopware-cli/logging"
 )
 
@@ -372,11 +373,11 @@ func init() {
 }
 
 func projectCISafetyCheck(ctx context.Context, root string, force bool, getenv func(string) string) error {
-	if force || isCIEnvironment(getenv) {
+	if force || system.IsCIEnvironment(getenv) {
 		return nil
 	}
 
-	dirty, isGitRepository, err := isGitWorkingTreeDirty(ctx, root)
+	dirty, isGitRepository, err := internalgit.IsWorkingTreeDirty(ctx, root)
 	if err != nil {
 		return err
 	}
@@ -393,56 +394,6 @@ func projectCISafetyCheck(ctx context.Context, root string, force bool, getenv f
 	logging.FromContext(ctx).Warnf("Running project ci outside a CI environment; this command removes source files and should usually only be used in CI")
 
 	return nil
-}
-
-func isCIEnvironment(getenv func(string) string) bool {
-	if strings.EqualFold(getenv("CI"), "true") {
-		return true
-	}
-
-	ciEnvVars := []string{
-		"GITHUB_ACTIONS",
-		"GITLAB_CI",
-		"JENKINS_URL",
-		"BUILDKITE",
-		"CIRCLECI",
-		"DRONE",
-		"TEAMCITY_VERSION",
-		"TF_BUILD",
-	}
-
-	for _, envVar := range ciEnvVars {
-		if getenv(envVar) != "" {
-			return true
-		}
-	}
-
-	return false
-}
-
-func isGitWorkingTreeDirty(ctx context.Context, root string) (bool, bool, error) {
-	cmd := exec.CommandContext(ctx, "git", "-C", root, "rev-parse", "--is-inside-work-tree") //nolint:gosec
-	output, err := cmd.Output()
-	if err != nil {
-		var exitErr *exec.ExitError
-		if errors.As(err, &exitErr) {
-			return false, false, nil
-		}
-
-		return false, false, fmt.Errorf("checking git repository: %w", err)
-	}
-
-	if strings.TrimSpace(string(output)) != "true" {
-		return false, false, nil
-	}
-
-	statusCmd := exec.CommandContext(ctx, "git", "-C", root, "status", "--porcelain", "--untracked-files=all") //nolint:gosec
-	status, err := statusCmd.Output()
-	if err != nil {
-		return false, true, fmt.Errorf("checking git working tree status: %w", err)
-	}
-
-	return strings.TrimSpace(string(status)) != "", true, nil
 }
 
 func commandWithRoot(cmd *exec.Cmd, root string) *exec.Cmd {
